move merkle to base
